change message id type to uint32_t

+++ src/carrier/ela_carrier_impl.h
@@ -96,7 +96,7 @@ typedef struct OfflineEvent {
 typedef struct MsgidEvent {
     EventBase base;
     char friendid[ELA_MAX_ADDRESS_LEN + 1];
-    int64_t msgid;
+    uint32_t msgid;
     int errcode;
 } MsgidEvent;
